changed problem with asynchronus
