+++ src/index.html
@@ -33,5 +33,6 @@
     <load src="./partials/mobile-menu.html" />
 
     <script type="module" src="main.js"></script>
+    <script type="module" src="./js/reviews.js"></script>
   </body>
 </html>
